+++ Tests/MapboxMapsTests/Foundation/Camera/BasicCameraAnimatorTests.swift
@@ -95,6 +95,152 @@ final class BasicCameraAnimatorTests: XCTestCase {
         XCTAssertTrue(delegate.cameraAnimatorDidStopRunningStub.parameters.first === animator)
     }
 
+    func testStartAndStopAnimationAfterDelay() throws {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        let randomInterval: TimeInterval = .random(in: 0...10)
+        animator.startAnimation(afterDelay: randomInterval)
+
+        XCTAssertEqual(self.propertyAnimator.startAnimationStub.invocations.count, 1)
+        XCTAssertEqual(self.propertyAnimator.addAnimationsStub.invocations.count, 1)
+        XCTAssertEqual(self.propertyAnimator.addCompletionStub.invocations.count, 1)
+
+        self.animator.stopAnimation()
+
+        XCTAssertEqual(self.propertyAnimator.stopAnimationStub.invocations.count, 1)
+        XCTAssertEqual(self.propertyAnimator.finishAnimationStub.invocations.count, 1)
+        XCTAssertEqual(self.propertyAnimator.finishAnimationStub.invocations.first?.parameters, .current)
+    }
+
+    func testCompletionBlockCalledForStartAndStopAfterDelay() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+
+        let expectation = XCTestExpectation(description: "The completion for the animator should be called when the animation is stopped.")
+
+        let completion: AnimationCompletion = { _ in
+            expectation.fulfill()
+        }
+        animator.addCompletion(completion)
+
+        let randomInterval: TimeInterval = .random(in: 1...10)
+        animator.startAnimation(afterDelay: randomInterval)
+
+        animator.stopAnimation()
+
+        wait(for: [expectation], timeout: 15)
+    }
+
+    func testStartandPauseAnimationAfterDelay() throws {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+
+        animator.startAnimation(afterDelay: 1)
+
+        animator.pauseAnimation()
+
+        XCTAssertEqual(propertyAnimator.pauseAnimationStub.invocations.count, 1)
+        XCTAssertEqual(propertyAnimator.stopAnimationStub.invocations.count, 0)
+    }
+
+    func testStartAnimationAfterDelayIsRunning() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        XCTAssertFalse(propertyAnimator.isRunning)
+
+        animator.startAnimation(afterDelay: 1)
+        XCTAssertTrue(animator.isRunning)
+    }
+
+    func testStartAndPauseAnimationAfterDelayIsNotRunning() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        XCTAssertFalse(propertyAnimator.isRunning)
+
+        animator.startAnimation(afterDelay: 1)
+        XCTAssertTrue(animator.isRunning)
+
+        animator.pauseAnimation()
+        XCTAssertFalse(animator.isRunning)
+    }
+
+    func testStartAndStopAnimationAfterDelayIsNotRunning() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        XCTAssertFalse(propertyAnimator.isRunning)
+
+        animator.startAnimation(afterDelay: 1)
+        XCTAssertTrue(animator.isRunning)
+
+        animator.stopAnimation()
+        XCTAssertFalse(propertyAnimator.isRunning)
+        XCTAssertFalse(animator.isRunning)
+    }
+
+    func testStartAnimationAfterDelayStateActive() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        XCTAssertEqual(animator.state, .inactive, "The animator's state should be inactive (0) since it hasn't started. Got \(animator.state.NSNumber).")
+
+        animator.startAnimation(afterDelay: 1)
+        XCTAssertEqual(animator.state, .active, "The animator's state should be active (1) since it is delayed. Got \(animator.state.NSNumber).")
+    }
+
+    func testStartAndPauseAnimationAfterDelayStateActive() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        animator.startAnimation(afterDelay: 1)
+        animator.pauseAnimation()
+        XCTAssertEqual(propertyAnimator.state, .active)
+        XCTAssertEqual(animator.state, .active, "The animator's state should be active (1). Got \(animator.state.rawValue).")
+    }
+
+    func testStartAndStopAnimationAfterDelayStateInactive() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        animator.startAnimation(afterDelay: 1)
+        animator.stopAnimation()
+        XCTAssertEqual(propertyAnimator.state, .inactive)
+        XCTAssertEqual(animator.state, .inactive, "The animator's state should be inactive (0). Got \(animator.state.rawValue).")
+    }
+
+    func testStartAnimationAfterDelayTransitionNotNil() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        animator.startAnimation(afterDelay: 1)
+        XCTAssertNotNil(animator.transition, "The transition should not be nil while the animation is delayed. Got \(animator.transition.debugDescription)")
+    }
+
+    func testPauseAndStartAnimationAfterDelayError() throws {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        animator.pauseAnimation()
+        expectFatalError(expectedMessage: "startAnimation(afterDelay:) cannot be called on already-delayed, paused, running, or completed animators.") {
+            self.animator.startAnimation(afterDelay: 0)
+        }
+    }
+
+    func testStartandPauseAnimationAfterDelayTransitionNotNil() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraStateTestValue.zoom
+        }
+        animator.startAnimation(afterDelay: 1)
+        animator.pauseAnimation()
+
+        XCTAssertNotNil(animator.transition, "The animator's transition property should not be nil after pausing the animation.")
+    }
+
     func testInformsDelegateWhenPausingAndStarting() {
         animator.addAnimations { (transition) in
             transition.zoom.toValue = cameraOptionsTestValue.zoom!
@@ -103,6 +249,15 @@ final class BasicCameraAnimatorTests: XCTestCase {
         XCTAssertEqual(delegate.cameraAnimatorDidStartRunningStub.invocations.count, 0)
 
         animator.startAnimation()
+        XCTAssertEqual(delegate.cameraAnimatorDidStartRunningStub.invocations.count, 1)
+        XCTAssertTrue(delegate.cameraAnimatorDidStartRunningStub.parameters.first === animator)
+    }
+
+    func testInformsDelegateWhenStartingAfterDelay() {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraOptionsTestValue.zoom!
+        }
+        animator.startAnimation(afterDelay: 1)
         XCTAssertEqual(delegate.cameraAnimatorDidStartRunningStub.invocations.count, 1)
         XCTAssertTrue(delegate.cameraAnimatorDidStartRunningStub.parameters.first === animator)
     }